requirements.txt
pandas>=2.0.0
python-docx>=1.0.0
pyinstaller>=6.0.0
tkinterdnd2>=0.3.0
